requirements.txt
ultralytics==8.3.59
numpy==1.26.4
RPi.GPIO==0.7.1
gpiozero==1.6.2
paho-mqtt==1.6.1
